% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caption.R
\name{get_caption}
\alias{get_caption}
\title{Retrieve the caption of an object}
\usage{
get_caption(id)
}
\arguments{
\item{id}{ID of the object (table or figure).}
}
\description{
Given the ID of an object, retrieve the corresponding caption in the ToT (table of tables), that must be located in the \code{Rdata/tot.rds} file.
}
