% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tot.R
\name{import_tot}
\alias{import_tot}
\title{Import the Table of Tables (ToT) from an Excel file}
\usage{
import_tot(
  tot_xlsx_path = here::here("data/tot/table-of-tables.xlsx"),
  verbose = FALSE
)
}
\arguments{
\item{tot_xlsx_path}{A character string specifying the path to the Excel file
containing the Table of Tables. Default is the file located at
"data/tot/table-of-tables.xlsx".}

\item{verbose}{A logical indicating whether to print verbose output during
the import process. Default is FALSE.}
}
\value{
None. The function is used for its side effects: importing,
validating, and saving the ToT data, and optionally printing verbose
output.
}
\description{
This function imports a Table of Tables (ToT) from an Excel file, validates
the data, and optionally provides verbose output about the import process.
The function performs multiple checks on the data, which are described below.
The imported data is saved as an RDS file.
}
\details{
The function performs the following checks on the ToT:
\itemize{
\item Required columns ("id", "name", "type", "orientation", "caption")
are present.
\item There are no duplicate names or IDs.
\item Names only contain lowercase and uppercase letters (a-z, A-Z),
digits (0-9), and dashes (-).
\item The \code{type} column contains only "figure" or "table".
\item The \code{orientation} column contains only "landscape" or "portrait".
}
}
\examples{
\dontrun{
# Example usage with default path and verbose output
import_tot(verbose = TRUE)

# Example usage with a specified file path
import_tot(tot_xlsx_path = "path/to/your/table-of-tables.xlsx", verbose = TRUE)
}

}
