% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_figure.R
\name{use_figure}
\alias{use_figure}
\title{Generate a script to create a figure}
\usage{
use_figure(name, id, author = "Alexandre Bohyn")
}
\arguments{
\item{name}{Name of the figure to create.}

\item{id}{ID value used to get the caption of the figure. The caption will be
obtained from the ToT}

\item{author}{Author of the script (defaults to the package creator for
simplicity)}
}
\description{
Create a \code{R} script to generate a figure in a report, using the template from
the \link{lbstProject} package
}
