% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_table.R
\name{use_table}
\alias{use_table}
\title{Generate a script to create a table}
\usage{
use_table(name, id, author = "Alexandre Bohyn")
}
\arguments{
\item{name}{Name of the table to create}

\item{id}{ID value used to get the caption of the table The caption will be
obtained from the ToT}

\item{author}{Author of the script (defaults to the package creator for
simplicity)}
}
\description{
Create a \code{R} script to generate a table in a report, using the template from
the \link{lbstProject} package
}
