% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose_ggsave.R
\name{verbose_ggsave}
\alias{verbose_ggsave}
\title{Save a plot with verbose}
\usage{
verbose_ggsave(plot, name, verbose = TRUE, ...)
}
\arguments{
\item{name}{Name of the plot that will be used to name the files generated
while saving the plot object and its data.}

\item{verbose}{Boolean. Display information about the path of the saved
files. Default is TRUE.}

\item{...}{Other arguments passed on to the \link[ggplot2:ggsave]{ggsave} function.}
}
\description{
Save a plot object as a png file and as a rds object with a verbose output
specifying the path of the saved files. The plot is saved in the
\code{results/figures} folder using \link[ggplot2:ggsave]{ggsave}, and the data file of the
plot is saved in the \code{Rdata/figures} folder using \link[base:readRDS]{saveRDS}.
}
