% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose_gtsave.R
\name{verbose_gtsave}
\alias{verbose_gtsave}
\title{Save a gt table with verbose}
\usage{
verbose_gtsave(data, name, extension = "docx", verbose = TRUE, ...)
}
\arguments{
\item{data}{The gt table data object, commonly created through the use of the
\code{\link[=gt]{gt()}} function.}

\item{name}{Name of the plot that will be used to name the files generated
while saving the plot object and its data.}

\item{extension}{Extension defining the output to which the table will be
saved. Accepted extensions are \code{.html}, \code{.tex}, \code{.ltx}, \code{.rtf}, \code{.docx}.}

\item{verbose}{Boolean. Display information about the path of the saved
files. Default is TRUE.}

\item{...}{Other arguments passed on to the \link[gt:gtsave]{gtsave} function.}
}
\description{
Save a gt table (created with the \link{gt} package) as a single docx file, and
also as a rds object with a verbose output specifying the path of the saved
files. The plot is saved in the \code{results/tables} folder using
\link[gt:gtsave]{gtsave}, and the data file of the plot is saved in the
\code{Rdata/tables} folder using \link[base:readRDS]{saveRDS}.
}
